update fee_state in `validate_proposal`
  * move state related types to state.rs
  * add fee state validation
  * update test
