+++ sequencer/src/state.rs
@@ -1,15 +1,25 @@
-use crate::{
-    block::{BlockMerkleTree, FeeAccount, FeeAmount, FeeMerkleTree},
-    Header, NodeState, Payload,
+use std::ops::Add;
+
+use crate::{Header, NodeState, Payload};
+use ark_serialize::{
+    CanonicalDeserialize, CanonicalSerialize, Compress, Read, SerializationError, Valid, Validate,
 };
+
+use derive_more::Add;
+use ethers::{abi::Address, types::U256};
+
 use commit::{Commitment, Committable};
 use hotshot::traits::ValidatedState as HotShotState;
 use hotshot_types::data::{BlockError, ViewNumber};
 use hotshot_types::traits::block_contents::BlockHeader;
+use jf_primitives::merkle_tree::prelude::{LightWeightSHA3MerkleTree, Sha3Node};
 use jf_primitives::merkle_tree::{
-    AppendableMerkleTreeScheme, ForgetableMerkleTreeScheme, MerkleTreeScheme,
+    prelude::Sha3Digest, universal_merkle_tree::UniversalMerkleTree, AppendableMerkleTreeScheme,
+    ForgetableMerkleTreeScheme, LookupResult, MerkleTreeScheme,
 };
+use jf_primitives::merkle_tree::{ToTraversalPath, UniversalMerkleTreeScheme};
 use serde::{Deserialize, Serialize};
+use typenum::Unsigned;
 
 #[derive(Hash, Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
 pub struct ValidatedState {
@@ -48,7 +58,7 @@ impl ValidatedState {
             )
         );
 
-        let fee_merkle_tree = self.fee_merkle_tree.clone();
+        let mut fee_merkle_tree = self.fee_merkle_tree.clone();
         let mut block_merkle_tree = self.block_merkle_tree.clone();
 
         // validate proposal is descendent of parent by appending to parent
@@ -57,9 +67,33 @@ impl ValidatedState {
         anyhow::ensure!(
             proposal.block_merkle_tree_root == block_merkle_tree_root,
             anyhow::anyhow!(
-                "Invalid Root Error: {}, {}",
+                "Invalid Block Root Error: {}, {}",
                 block_merkle_tree_root,
                 proposal.block_merkle_tree_root
+            )
+        );
+
+        // fetch receipts from the l1
+        let receipts = fetch_fee_receipts(parent);
+        for FeeReceipt { recipient, amount } in receipts {
+            // Get the balance in order to add amount, ignoring the proof.
+            match fee_merkle_tree.universal_lookup(recipient) {
+                LookupResult::Ok(balance, _) => fee_merkle_tree
+                    .update(recipient, balance.add(amount))
+                    .unwrap(),
+                // Handle `NotFound` and `NotInMemory` by initializing
+                // state.
+                _ => fee_merkle_tree.update(recipient, amount).unwrap(),
+            };
+        }
+
+        let fee_merkle_tree_root = fee_merkle_tree.commitment();
+        anyhow::ensure!(
+            proposal.fee_merkle_tree_root == fee_merkle_tree_root,
+            anyhow::anyhow!(
+                "Invalid Fee Root Error: {}, {}",
+                fee_merkle_tree_root,
+                proposal.fee_merkle_tree_root
             )
         );
 
@@ -132,3 +166,112 @@ impl hotshot_types::traits::states::TestableState for ValidatedState {
         crate::Transaction::random(rng)
     }
 }
+
+pub type BlockMerkleTree = LightWeightSHA3MerkleTree<Commitment<Header>>;
+pub type BlockMerkleCommitment = <BlockMerkleTree as MerkleTreeScheme>::Commitment;
+
+// New Type for `U256` in order to implement `CanonicalSerialize` and
+// `CanonicalDeserialize`
+#[derive(Default, Hash, Copy, Clone, Debug, Deserialize, Serialize, PartialEq, Eq, Add)]
+pub struct FeeAmount(U256);
+// New Type for `Address` in order to implement `CanonicalSerialize` and
+// `CanonicalDeserialize`
+#[derive(
+    Default, Hash, Copy, Clone, Debug, Deserialize, Serialize, PartialEq, Eq, PartialOrd, Ord,
+)]
+pub struct FeeAccount(Address);
+
+impl Valid for FeeAmount {
+    fn check(&self) -> Result<(), SerializationError> {
+        Ok(())
+    }
+}
+
+impl Valid for FeeAccount {
+    fn check(&self) -> Result<(), SerializationError> {
+        Ok(())
+    }
+}
+
+impl CanonicalSerialize for FeeAmount {
+    fn serialize_with_mode<W: std::io::prelude::Write>(
+        &self,
+        mut writer: W,
+        _compress: Compress,
+    ) -> Result<(), SerializationError> {
+        let mut bytes = [0u8; core::mem::size_of::<U256>()];
+        self.0.to_little_endian(&mut bytes);
+        Ok(writer.write_all(&bytes)?)
+    }
+
+    fn serialized_size(&self, _compress: Compress) -> usize {
+        core::mem::size_of::<U256>()
+    }
+}
+impl CanonicalDeserialize for FeeAmount {
+    fn deserialize_with_mode<R: Read>(
+        mut reader: R,
+        _compress: Compress,
+        _validate: Validate,
+    ) -> Result<Self, SerializationError> {
+        let mut bytes = [0u8; core::mem::size_of::<U256>()];
+        reader.read_exact(&mut bytes)?;
+        let value = U256::from_little_endian(&bytes);
+        Ok(Self(value))
+    }
+}
+impl CanonicalSerialize for FeeAccount {
+    fn serialize_with_mode<W: std::io::prelude::Write>(
+        &self,
+        mut writer: W,
+        _compress: Compress,
+    ) -> Result<(), SerializationError> {
+        Ok(writer.write_all(self.0.as_bytes())?)
+    }
+
+    fn serialized_size(&self, _compress: Compress) -> usize {
+        core::mem::size_of::<Address>()
+    }
+}
+impl CanonicalDeserialize for FeeAccount {
+    fn deserialize_with_mode<R: Read>(
+        mut reader: R,
+        _compress: Compress,
+        _validate: Validate,
+    ) -> Result<Self, SerializationError> {
+        let mut bytes = [0u8; core::mem::size_of::<Address>()];
+        reader.read_exact(&mut bytes)?;
+        let value = Address::from_slice(&bytes);
+        Ok(Self(value))
+    }
+}
+impl std::convert::From<u64> for FeeAccount {
+    fn from(item: u64) -> Self {
+        FeeAccount(Address::from_low_u64_le(item))
+    }
+}
+
+impl<A: Unsigned> ToTraversalPath<A> for FeeAccount {
+    fn to_traversal_path(&self, height: usize) -> Vec<usize> {
+        Address::to_fixed_bytes(self.0)
+            .into_iter()
+            .take(height)
+            .map(|i| i as usize)
+            .collect()
+    }
+}
+
+#[derive(Default, Hash, Clone, CanonicalDeserialize)]
+pub struct FeeReceipt {
+    pub recipient: FeeAccount,
+    pub amount: FeeAmount,
+}
+/// Fetch fee receitps from l1. Currently a mock function to be
+/// implemented in the future.
+pub fn fetch_fee_receipts(_parent: &Header) -> Vec<FeeReceipt> {
+    Vec::from([FeeReceipt::default()])
+}
+
+pub type FeeMerkleTree =
+    UniversalMerkleTree<FeeAmount, Sha3Digest, FeeAccount, typenum::U256, Sha3Node>;
+pub type FeeMerkleCommitment = <FeeMerkleTree as MerkleTreeScheme>::Commitment;
